+++ src/SchemaGenerator.ts
@@ -13,6 +13,8 @@ import { removeUnreachable } from "./Utils/removeUnreachable";
 import { Config } from "./Config";
 import { hasJsDocTag } from "./Utils/hasJsDocTag";
 import { JSONSchema7, JSONSchema7Definition } from "json-schema";
+import { unambiguousName } from "./Utils/unambiguousName";
+import { resolveIdRefs } from "./Utils/resolveIdRefs";
 
 export class SchemaGenerator {
     public constructor(
@@ -176,117 +178,4 @@ export class SchemaGenerator {
         const symbol = symbolAtNode(node)!;
         return typeChecker.getFullyQualifiedName(symbol).replace(/".*"\./, "");
     }
-}
-
-/**
- * Given a set of paths, will strip the common-prefix, and return an array of remaining substrings.
- * Also removes any file extensions, and replaces any '/' with '-' in the path.
- * Each return value can be used as a name prefix for disambiguation.
- * The returned prefixes array maintains input order.
- */
-function getCommonPrefixes(paths: string[]) {
-    // clone before sorting to maintain input order.
-    const sorted = [...paths].sort();
-    const shortest = sorted[0];
-    const second = sorted[sorted.length - 1];
-    const maxPrefix = shortest.length;
-    let pos = 0;
-    let path_pos = 0;
-    while (pos < maxPrefix && shortest.charAt(pos) === second.charAt(pos)) {
-        if (shortest.charAt(pos) === "/") {
-            path_pos = pos;
-        }
-        pos++;
-    }
-    return paths.map((p) =>
-        p
-            .substr(path_pos + 1)
-            .replace(/\//g, "__")
-            .replace(/\.[^.]+$/, "")
-    );
-}
-
-function unambiguousName(child: DefinitionType, isRoot: boolean, peers: DefinitionType[]): string {
-    if (peers.length === 1 || isRoot) {
-        return child.getName();
-    } else if (child.getType().getSrcFileName()) {
-        let index = -1;
-
-        // filter unique peers to be those that have srcFileNames.
-        // Intermediate Types - AnnotationTypes, UnionTypes, do not have sourceFileNames
-        const uniques = peers.filter((peer) => peer.getType().getSrcFileName());
-        if (uniques.length === 1) {
-            return uniques[0].getName();
-        }
-        const srcPaths = uniques.map((peer: DefinitionType, count) => {
-            index = child === peer ? count : index;
-            return peer.getType().getSrcFileName()!;
-        });
-        const prefixes = getCommonPrefixes(srcPaths);
-        return `${prefixes[index]}-${child.getName()}`;
-    } else {
-        // intermediate type.
-        const name = child.getName();
-        // TODO: Perhaps we should maintain a name-id map, and throw a duplicate error on collision.
-        if (name === undefined) {
-            // this might be unreachable code.
-            throw new Error(`Unable to disambiguate types`);
-        }
-        return name;
-    }
-}
-
-/**
- * Resolve all `#/definitions/...` and `$ref` in schema with appropriate disambiguated names
- */
-function resolveIdRefs(
-    schema: JSONSchema7Definition,
-    idNameMap: Map<string, string>,
-    encodeRefs: boolean
-): JSONSchema7Definition {
-    if (!schema || typeof schema === "boolean") {
-        return schema;
-    }
-    const { $ref, allOf, oneOf, anyOf, not, properties, items, definitions, additionalProperties, ...rest } = schema;
-    const result: JSONSchema7Definition = { ...rest };
-    if ($ref) {
-        // THE Money Shot.
-        const id = encodeRefs ? decodeURIComponent($ref.slice(14)) : $ref.slice(14);
-        const name = idNameMap.get(id);
-        result.$ref = `#/definitions/${encodeRefs ? encodeURIComponent(name!) : name}`;
-    }
-    if (definitions) {
-        result.definitions = Object.entries(definitions).reduce((acc, [prop, value]) => {
-            const name = idNameMap.get(prop)!;
-            acc[name] = resolveIdRefs(value, idNameMap, encodeRefs);
-            return acc;
-        }, {} as StringMap<JSONSchema7Definition>);
-    }
-    if (properties) {
-        result.properties = Object.entries(properties).reduce((acc, [prop, value]) => {
-            acc[prop] = resolveIdRefs(value, idNameMap, encodeRefs);
-            return acc;
-        }, {} as StringMap<JSONSchema7Definition>);
-    }
-    if (additionalProperties || additionalProperties === false) {
-        result.additionalProperties = resolveIdRefs(additionalProperties, idNameMap, encodeRefs);
-    }
-    if (items) {
-        result.items = Array.isArray(items)
-            ? items.map((el) => resolveIdRefs(el, idNameMap, encodeRefs))
-            : resolveIdRefs(items, idNameMap, encodeRefs);
-    }
-    if (allOf) {
-        result.allOf = allOf.map((el) => resolveIdRefs(el, idNameMap, encodeRefs));
-    }
-    if (anyOf) {
-        result.anyOf = anyOf.map((el) => resolveIdRefs(el, idNameMap, encodeRefs));
-    }
-    if (oneOf) {
-        result.oneOf = oneOf.map((el) => resolveIdRefs(el, idNameMap, encodeRefs));
-    }
-    if (not) {
-        result.not = resolveIdRefs(not, idNameMap, encodeRefs);
-    }
-    return result;
 }
